"Bulletproof" sendData method, +SoftwareSerial
Created a sendData method that can be passed raw bytes to send over the
XBee. You should be able to pass it anything as all bytes are properly
escaped. This function does a lot of undocumented stuff.

For instance, only pre-escaped bytes are used to calculate the
checksum. Good luck finding that in the documentation. Also, the entire
frame can be escaped. This includes the length fields but not the
initial 0x7E. I’m still uncertain if this includes the checksum or not.
Right now I’m not escaping the checksum.

Added SoftwareSerial support by abstracting the methods I needed and
checking which pointer (HardwareSerial or SoftwareSerial) it was
initialized with.